Add `epoch_slot` to block receipt delay.
This adds a label `epoch_slot` to the prometheus metric
`vouch_block_receipt_delay_seconds`.  It has been noted that the receipt
delay is often significantly higher for the first (and, to a lesser
extent, second) slot in a given epoch.  This allows the receipt delay to
be examined for specific slots in a given epoch.