Do not overwrite validator list with null.
If Vouch contains a list of validators, and on refresh obtains an empty
list, assume the empty response is a result of an error and do not
remove the existing list in favor of it.

It is possible for network issues to result in an empty response when
requesting an update on the list of validators for which Vouch should
operate.  In this situation, where Vouch already has a list of
validators, it retains that list rather than replace it with the empty
results of the request.