Add advanced scheduler.
The basic scheduler has shown weakness when handling high load, which
manifests with the occasional deadlock situation.  This introduces an
advanced scheduler, which contains additional features in an attempt to
ensure that regardless of the concurrent load no job is ever left in an
inconsistent state, explicitly with outstanding items to read on its
channels.