Fetch next epoch duties at an appropriate time.
Duties for the subsequent epoch were being fetched at the turn of the
epoch prior.  This resulted in high load on the beacon node for the
first slot of the epoch.

This change schedules the preparation for the next epoch to be at a
later time in the current epoch, spreading out the load and avoiding
overloading the beacon node at its busiest time.