Avoid deadlock in advanced scheduler.
It was possible for a job to be triggered just before it was due to
run due to its timer.  This situation would cause a deadlock, resulting
in Vouch terminating itself.  This adds explicit early removal from the
jobs list as well as additional checks when trigger conditions are met.