Controlled shutdown.
This changes Vouch's behaviour on a shutdown request, to wait for any
outstanding attestation process for the current slot before starting the
shutdown procedure.

Note that if there is an outstanding block proposal it should complete
prior to the attestation starting, so is covered by the same process.
If there is an outstanding sync committee duty we expect it to complete
prior to the attestation, but even if not the impact of its loss will be
relatively small compared to that of the attestations.