Ensure that registration gas limit is not 0.
It was possible for a registration's gas limit to be 0 if not present if
the default configuration of a V1 proposer configuration did not provide
a gas limit and it was called upon to be used in lieu of a specific
proposer configuration.  Ensure that the fallback gas limit is used in
this situation.