feat: add blockrelay.privileged-builders
Privileged builders are a list of public keys of builders from which bids
will be accepted first, even if the bid is lower. If no bids are
received from the priority builders, bids from other builders will be
accepted.

This is useful when you have a special contract with specific relays and
you would like to have their blocks used instead of other relays. While
still having other relays as a fallback.