Use correct mime type on fetch of execution config
When vouch was configured to fetch the blockrelay/executionconfig from
an http endpoint, the respective majordomo call did not properly set the
`Content-Type` and `Accept` headers. Instead, `Content-Type: text/plain`
was sent.

This commit adds the respective `MIMEType` value to the majordomo
context so that these requests are now sent with the `Content-Type` and
`Accept` headers set to `application/json`.